Added control flow in the form of async, logging in the form of debug and helpers in the form of lodash.
Moved code I used to parse stop loss into a function, safely parse
number.  Replaced the while with async.doWhilst because I needed a
while loop with a callback.

Implemented part of the stop loss bot, it now checks the sell price,
compares that to the stop loss price and if the sell price is below the
stop loss price it will exit else it goes again.